+++ packages/instagram/CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies [b34aa2e]
-  - @microfox/instagram-business-oauth@1.2.0
+  - @microfox/instagram-oauth@1.2.0
 
 ## 1.2.0
 
@@ -20,7 +20,7 @@
 ### Patch Changes
 
 - Updated dependencies [16d112d]
-  - @microfox/instagram-business-oauth@1.1.0
+  - @microfox/instagram-oauth@1.1.0
 
 ## 1.1.1
 
@@ -28,7 +28,7 @@
 
 - 7cd211d: Changes from PR #251: brave
 - Updated dependencies [7cd211d]
-  - @microfox/instagram-business-oauth@1.0.2
+  - @microfox/instagram-oauth@1.0.2
 
 ## 1.1.0